Simulator/requirements.txt
pygame 
python-can